requirements.txt
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
contourpy==1.2.1
cycler==0.12.1
dill==0.3.8
fonttools==4.51.0
gseapy==1.1.2
idna==3.7
joblib==1.4.2
kiwisolver==1.4.5
matplotlib==3.8.4
nltk==3.8.1
numpy==1.26.4
packaging==24.0
pandas==2.2.2
pillow==10.3.0
pyparsing==3.1.2
python-dateutil==2.9.0.post0
pytz==2024.1
regex==2024.5.10
requests==2.31.0
scipy==1.13.0
six==1.16.0
tqdm==4.66.4
tzdata==2024.1
urllib3==2.2.1
wordcloud==1.9.3
